Move AsyncRedis class to the utils module
